# contracts/openapi.json
{
  "components": {
    "schemas": {
      "CommandRecord": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "request": {
            "$ref": "#/components/schemas/CommandRequest"
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus"
          },
          "transitions": {
            "items": {
              "$ref": "#/components/schemas/CommandTransition"
            },
            "title": "Transitions",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "request",
          "status",
          "transitions",
          "updated_at"
        ],
        "title": "CommandRecord",
        "type": "object"
      },
      "CommandRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "discriminator": {
              "mapping": {
                "manipulate": "#/components/schemas/ManipulateAction",
                "navigate": "#/components/schemas/NavigateAction",
                "protective_stop": "#/components/schemas/ProtectiveStopAction"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/NavigateAction"
              },
              {
                "$ref": "#/components/schemas/ManipulateAction"
              },
              {
                "$ref": "#/components/schemas/ProtectiveStopAction"
              }
            ],
            "title": "Action"
          },
          "command_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Command Id",
            "type": "string"
          },
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "expected_state_version": {
            "minimum": 0.0,
            "title": "Expected State Version",
            "type": "integer"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "robot_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Robot Id",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "robot_id",
          "issued_at",
          "expires_at",
          "expected_state_version",
          "action"
        ],
        "title": "CommandRequest",
        "type": "object"
      },
      "CommandStatus": {
        "enum": [
          "submitted",
          "accepted",
          "rejected",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "CommandStatus",
        "type": "string"
      },
      "CommandTransition": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus"
          }
        },
        "required": [
          "status",
          "occurred_at"
        ],
        "title": "CommandTransition",
        "type": "object"
      },
      "CommandType": {
        "enum": [
          "navigate",
          "manipulate",
          "protective_stop"
        ],
        "title": "CommandType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HardwareSafetyState": {
        "enum": [
          "normal",
          "estop_engaged"
        ],
        "title": "HardwareSafetyState",
        "type": "string"
      },
      "HardwareStatePatch": {
        "additionalProperties": false,
        "description": "Test/simulator-only hardware observation, never a software E-stop command.",
        "properties": {
          "hardware_estop_engaged": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hardware Estop Engaged"
          },
          "joint_positions_rad": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "maxItems": 16,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joint Positions Rad"
          },
          "pose": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Pose2D"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "HardwareStatePatch",
        "type": "object"
      },
      "JointLimit": {
        "additionalProperties": false,
        "description": "A mock product's planner-visible position and velocity envelope.",
        "properties": {
          "joint_name": {
            "pattern": "^[A-Za-z][A-Za-z0-9_]{0,63}$",
            "title": "Joint Name",
            "type": "string"
          },
          "max_position_rad": {
            "maximum": 6.283186,
            "minimum": -6.283186,
            "title": "Max Position Rad",
            "type": "number"
          },
          "max_velocity_rad_s": {
            "exclusiveMinimum": 0.0,
            "maximum": 20.0,
            "title": "Max Velocity Rad S",
            "type": "number"
          },
          "min_position_rad": {
            "maximum": 6.283186,
            "minimum": -6.283186,
            "title": "Min Position Rad",
            "type": "number"
          }
        },
        "required": [
          "joint_name",
          "min_position_rad",
          "max_position_rad",
          "max_velocity_rad_s"
        ],
        "title": "JointLimit",
        "type": "object"
      },
      "ManipulateAction": {
        "additionalProperties": false,
        "properties": {
          "joint_positions_rad": {
            "items": {
              "type": "number"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Joint Positions Rad",
            "type": "array"
          },
          "max_force_n": {
            "default": 30,
            "exclusiveMinimum": 0.0,
            "maximum": 250.0,
            "title": "Max Force N",
            "type": "number"
          },
          "type": {
            "const": "manipulate",
            "default": "manipulate",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "joint_positions_rad"
        ],
        "title": "ManipulateAction",
        "type": "object"
      },
      "NavigateAction": {
        "additionalProperties": false,
        "properties": {
          "max_speed_mps": {
            "default": 0.5,
            "exclusiveMinimum": 0.0,
            "maximum": 2.0,
            "title": "Max Speed Mps",
            "type": "number"
          },
          "target": {
            "$ref": "#/components/schemas/Pose2D"
          },
          "type": {
            "const": "navigate",
            "default": "navigate",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "target"
        ],
        "title": "NavigateAction",
        "type": "object"
      },
      "Pose2D": {
        "additionalProperties": false,
        "properties": {
          "frame": {
            "default": "map",
            "pattern": "^[A-Za-z][A-Za-z0-9_/]{0,63}$",
            "title": "Frame",
            "type": "string"
          },
          "x_m": {
            "maximum": 10000.0,
            "minimum": -10000.0,
            "title": "X M",
            "type": "number"
          },
          "y_m": {
            "maximum": 10000.0,
            "minimum": -10000.0,
            "title": "Y M",
            "type": "number"
          },
          "yaw_rad": {
            "maximum": 3.141593,
            "minimum": -3.141593,
            "title": "Yaw Rad",
            "type": "number"
          }
        },
        "required": [
          "x_m",
          "y_m",
          "yaw_rad"
        ],
        "title": "Pose2D",
        "type": "object"
      },
      "ProductCapabilityProfile": {
        "additionalProperties": false,
        "description": "Strict v2 product envelope; the digest covers every preceding field.",
        "properties": {
          "base_frame": {
            "pattern": "^[A-Za-z][A-Za-z0-9_/]{0,63}$",
            "title": "Base Frame",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/CommandType"
            },
            "minItems": 1,
            "title": "Capabilities",
            "type": "array"
          },
          "classification": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Classification",
            "type": "string"
          },
          "connection_interface": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Connection Interface",
            "type": "string"
          },
          "joint_count": {
            "maximum": 16.0,
            "minimum": 1.0,
            "title": "Joint Count",
            "type": "integer"
          },
          "joint_limits": {
            "items": {
              "$ref": "#/components/schemas/JointLimit"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Joint Limits",
            "type": "array"
          },
          "max_manipulation_force_n": {
            "exclusiveMinimum": 0.0,
            "maximum": 250.0,
            "title": "Max Manipulation Force N",
            "type": "number"
          },
          "max_navigation_speed_mps": {
            "exclusiveMinimum": 0.0,
            "maximum": 2.0,
            "title": "Max Navigation Speed Mps",
            "type": "number"
          },
          "max_payload_kg": {
            "exclusiveMinimum": 0.0,
            "maximum": 1000.0,
            "title": "Max Payload Kg",
            "type": "number"
          },
          "model_name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "product_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Product Id",
            "type": "string"
          },
          "product_name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Product Name",
            "type": "string"
          },
          "profile_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Profile Digest",
            "type": "string"
          },
          "schema_version": {
            "const": "2.0.0",
            "default": "2.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "tool_frame": {
            "pattern": "^[A-Za-z][A-Za-z0-9_/]{0,63}$",
            "title": "Tool Frame",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "classification",
          "product_name",
          "model_name",
          "capabilities",
          "joint_count",
          "connection_interface",
          "joint_limits",
          "max_manipulation_force_n",
          "max_navigation_speed_mps",
          "max_payload_kg",
          "base_frame",
          "tool_frame",
          "profile_digest"
        ],
        "title": "ProductCapabilityProfile",
        "type": "object"
      },
      "ProductProfile": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/CommandType"
            },
            "title": "Capabilities",
            "type": "array",
            "uniqueItems": true
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "connection_interface": {
            "title": "Connection Interface",
            "type": "string"
          },
          "joint_count": {
            "maximum": 16.0,
            "minimum": 1.0,
            "title": "Joint Count",
            "type": "integer"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "classification",
          "product_name",
          "model_name",
          "capabilities",
          "joint_count",
          "connection_interface"
        ],
        "title": "ProductProfile",
        "type": "object"
      },
      "ProtectiveStopAction": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "type": {
            "const": "protective_stop",
            "default": "protective_stop",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ProtectiveStopAction",
        "type": "object"
      },
      "RobotState": {
        "additionalProperties": false,
        "properties": {
          "active_command_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Command Id"
          },
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "hardware_safety_state": {
            "$ref": "#/components/schemas/HardwareSafetyState"
          },
          "joint_positions_rad": {
            "items": {
              "type": "number"
            },
            "title": "Joint Positions Rad",
            "type": "array"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "pose": {
            "$ref": "#/components/schemas/Pose2D"
          },
          "product_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Product Id",
            "type": "string"
          },
          "robot_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
            "title": "Robot Id",
            "type": "string"
          },
          "software_protective_stop": {
            "title": "Software Protective Stop",
            "type": "boolean"
          },
          "state_version": {
            "minimum": 0.0,
            "title": "State Version",
            "type": "integer"
          }
        },
        "required": [
          "robot_id",
          "product_id",
          "state_version",
          "observed_at",
          "pose",
          "joint_positions_rad",
          "hardware_safety_state",
          "software_protective_stop"
        ],
        "title": "RobotState",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Vendor-neutral RobotPort. software protective_stop is not a certified hardware emergency stop.",
    "title": "Physical Robot Interface",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/commands": {
      "post": {
        "operationId": "submit_command_v1_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Command"
      }
    },
    "/v1/commands/{command_id}": {
      "get": {
        "operationId": "get_command_v1_commands__command_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "command_id",
            "required": true,
            "schema": {
              "title": "Command Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Command"
      }
    },
    "/v1/commands/{command_id}/cancel": {
      "post": {
        "operationId": "cancel_command_v1_commands__command_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "command_id",
            "required": true,
            "schema": {
              "title": "Command Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Command"
      }
    },
    "/v1/products": {
      "get": {
        "operationId": "list_products_v1_products_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductProfile"
                  },
                  "title": "Response List Products V1 Products Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Products"
      }
    },
    "/v1/robots": {
      "get": {
        "operationId": "list_robots_v1_robots_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RobotState"
                  },
                  "title": "Response List Robots V1 Robots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Robots"
      }
    },
    "/v1/robots/{robot_id}": {
      "get": {
        "operationId": "get_robot_v1_robots__robot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "robot_id",
            "required": true,
            "schema": {
              "title": "Robot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Robot"
      }
    },
    "/v1/simulator/robots/{robot_id}/hardware-state": {
      "patch": {
        "operationId": "simulate_hardware_state_v1_simulator_robots__robot_id__hardware_state_patch",
        "parameters": [
          {
            "in": "path",
            "name": "robot_id",
            "required": true,
            "schema": {
              "title": "Robot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HardwareStatePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Hardware State"
      }
    },
    "/v2/products": {
      "get": {
        "operationId": "list_product_capabilities_v2_products_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductCapabilityProfile"
                  },
                  "title": "Response List Product Capabilities V2 Products Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Product Capabilities"
      }
    },
    "/v2/products/{product_id}": {
      "get": {
        "operationId": "get_product_capability_v2_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCapabilityProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Product Capability"
      }
    }
  }
}